+++ src/widgets/container.rs
@@ -13,8 +13,8 @@ pub fn container<App>(child: Widget<App>) -> Container<App> {
 
 pub struct Container<App> {
     pub child: Widget<App>,
-    pub width: f32,
-    pub height: f32,
+    pub width: u32,
+    pub height: u32,
     pub color: Color,
     pub radius: f32,
     pub padding: Padding,
@@ -27,15 +27,15 @@ impl<App> Container<App> {
     pub fn new(child: Widget<App>) -> Self {
         Self {
             child,
-            width: 100.0,
-            height: 50.0,
+            width: 100,
+            height: 50,
             color: Color::rgb(50, 50, 51),
             radius: 0.0,
             padding: Padding {
-                top: 0.0,
-                left: 0.0,
-                right: 0.0,
-                bottom: 0.0,
+                top: 0,
+                left: 0,
+                right: 0,
+                bottom: 0,
             },
             // id: None,
             align: None,
@@ -44,7 +44,7 @@ impl<App> Container<App> {
         }
     }
 
-    pub fn size(mut self, width: f32, height: f32) -> Self {
+    pub fn size(mut self, width: u32, height: u32) -> Self {
         self.width = width;
         self.height = height;
         self
